+++ src/common/absolutePath.ts
@@ -1,9 +1,9 @@
-import * as path from "path";
+import * as path from 'path';
 
 export class AbsolutePath{
     constructor(public value: string) {
         if (!path.isAbsolute(value)) {
-            throw new Error("The path must be absolute");
+            throw new Error('The path must be absolute');
         }
     }