+++ src/extension.ts
@@ -1,15 +1,21 @@
 import * as vscode from 'vscode';
-import { PlatformInformation } from './common/PlatformInformation';
+import { CredentialController } from './credential/credentialController';
+import { uriHandler } from './shared';
+import { PlatformInformation } from './common/platformInformation';
 import { ensureRuntimeDependencies } from './dependency/dependencyManager';
-import { DocsLoggerObserver } from './observers/DocsLoggerObserver';
-import { DocsOutputChannelObserver } from './observers/DocsOutputChannelObserver';
-import ExtensionExports from './common/ExtensionExport';
-import { EventStream } from './common/EventStream';
-import { EnvironmentController } from './common/EnvironmentController';
+import { SignStatusBarObserver } from './observers/signStatusBarObserver';
+import { DocsLoggerObserver } from './observers/docsLoggerObserver';
+import { DocsOutputChannelObserver } from './observers/docsOutputChannelObserver';
+import { ErrorMessageObserver } from './observers/errorMessageObserver';
+import { InfoMessageObserver } from './observers/infoMessageObserver';
+import ExtensionExports from './common/extensionExport';
+import { EventStream } from './common/eventStream';
+import { KeyChain } from './credential/keyChain';
+import { DocsEnvironmentController } from './common/docsEnvironmentController';
 
 export async function activate(context: vscode.ExtensionContext): Promise<ExtensionExports> {
     const eventStream = new EventStream();
-    const environmentController = new EnvironmentController(eventStream);
+    const environmentController = new DocsEnvironmentController(eventStream);
     const platformInformation = await PlatformInformation.getCurrent();
 
     // Output Channel
@@ -21,25 +27,42 @@ export async function activate(context: vscode.ExtensionContext): Promise<Extens
 
     let runtimeDependenciesInstalled = await ensureRuntimeDependencies(platformInformation, eventStream);
     if (!runtimeDependenciesInstalled) {
-        throw new Error("Install runtime dependencies failed, Please restart Visual Studio Code to re-trigger the download.");
+        throw new Error('Install runtime dependencies failed, Please restart Visual Studio Code to re-trigger the download.');
     }
+
+    // Message 
+    let errorMessageObserver = new ErrorMessageObserver();
+    let infoMessageObserver = new InfoMessageObserver();
+    eventStream.subscribe(errorMessageObserver.eventHandler);
+    eventStream.subscribe(infoMessageObserver.eventHandler);
+
+    // Credential component initialize
+    let keyChain = new KeyChain(environmentController);
+    let credentialController = new CredentialController(keyChain, eventStream, environmentController);
+    eventStream.subscribe(credentialController.eventHandler);
+    // Initialize credential
+    let credentialInitialPromise = credentialController.initialize();
+
+    // Sign Status bar
+    let signStatusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, Number.MIN_VALUE + 1);
+    let signStatusBarObserver = new SignStatusBarObserver(signStatusBar, environmentController);
+    eventStream.subscribe(signStatusBarObserver.eventHandler);
 
     context.subscriptions.push(
         outputChannel,
         environmentController,
-        vscode.commands.registerCommand('docs.signIn', () => {
-            // TODO: handle sign in command
-        }),
-        vscode.commands.registerCommand('docs.signOut', () => {
-            // TODO: handle sign out command
-        }),
+        vscode.commands.registerCommand('docs.signIn', () => credentialController.signIn()),
+        vscode.commands.registerCommand('docs.signOut', () => credentialController.signOut()),
         vscode.commands.registerCommand('docs.build', async (uri) => {
             // TODO: handle build command
-        })
+        }),
+        vscode.window.registerUriHandler(uriHandler)
     );
 
     return {
-        initializationFinished: async () => { }
+        initializationFinished: async () => {
+            await credentialInitialPromise;
+        }
     };
 }